backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
google-generativeai>=0.3.2
python-pptx>=0.6.23
redis>=5.0.1
celery>=5.3.6
requests>=2.31.0
python-dotenv>=1.0.1
httpx>=0.26.0
